fix: remove existing timeframes from predefine timeframe
