+++ apps/gauzy/src/app/pages/goal-settings/goal-settings.component.html
@@ -62,7 +62,7 @@
 			</ng2-smart-table>
 			<ga-card-grid
 				[settings]="smartTableSettings"
-				[source]="goalSettingData"
+				[source]="goalTimeFrames"
 				*ngIf="dataLayoutStyle === 'CARDS_GRID'"
 				[buttonTemplate]="actionButtons"
 			></ga-card-grid>
