+++ apps/gauzy/src/assets/i18n/en.json
@@ -733,6 +733,11 @@
 			"PREDEFINED_TIME_FRAMES": "Predefined Time Frames",
 			"ADD_KPI": "Add KPI"
 		},
+		"MESSAGE": {
+			"NO_KEY_RESULT": "No Key Results to display.",
+			"NO_UPDATES": "No Updates yet.",
+			"NO_ALIGNMENT": "No Alignments yet."
+		},
 		"LEVELS": {
 			"ORGANIZATION": "Organization",
 			"TEAM": "Team",
